+++ include/Ark/VM/FFI.hpp
@@ -3,7 +3,6 @@
 
 #include <vector>
 #include <utility>
-#include <iostream>
 #include <sstream>
 
 #include <Ark/VM/Value.hpp>
@@ -34,6 +33,9 @@ namespace Ark::internal::FFI
     FFI_Function(readFile);    // readFile, 1 argument
     FFI_Function(fileExists);  // fileExists?, 1 argument
     FFI_Function(timeSinceEpoch);  // time, 0 argument
+    FFI_Function(sleep);  // sleep, 1 argument
+    FFI_Function(system_);  // system, 1 argument
+    FFI_Function(format);  // format, multiple arguments
 }
 
 #undef FFI_Function